+++ packages/snap-networks-utils/src/utils/logger/__mocks__/Logger.ts
@@ -1,4 +1,4 @@
-import type { Logger } from '../Logger';
+import type { Logger } from '../Logger.js';
 
 export const mockLogger = {
   log: jest.fn(),