+++ packages/snap-networks-utils/src/utils/sanitize/sanitize.test.ts
@@ -1,4 +1,4 @@
-import { sanitizeControlCharacters, sanitizeUri } from './sanitize';
+import { sanitizeControlCharacters, sanitizeUri } from './sanitize.js';
 
 describe('sanitize', () => {
   describe('sanitizeControlCharacters', () => {